create-empty-hashes.py - script to create empty blob and empty tree hash headers
The observation is that handling 10 hash algorithms requires
one to create generators. With 2, one can get away with doing
this stuff by hand.

Perhaps we could add a function that creates the empty blob and
empty tree oids. That would have the benefit of allowing us to
use a different scheme for hashes that were not susceptible to
length extension.

Signed-off-by: Michael Clark <michaeljclark@mac.com>