create-test-hashes.sh - script to create exemplar test hashes for t0015-hash.sh
The observation is that handling 10 hash algorithms requires
one to create generators. With 2, one can get away with doing
this stuff by hand.

This is a quick and dirty script. With some more work we could
create a generator that outputs the complete test cases of change
t015-hashes.sh to a loop that uses openssl as an out-of-band test.

Signed-off-by: Michael Clark <michaeljclark@mac.com>